관리자 페이지들 관리자만 접근할 수 있도록 수정

+++ Semi_Project/src/com/kh/admin/controller/DeleteBookServlet.java
@@ -10,6 +10,7 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 import com.kh.admin.model.service.AdminService;
+import com.kh.member.model.vo.Member;
 
 /**
  * Servlet implementation class DeleteBookServlet
@@ -31,33 +32,46 @@ public class DeleteBookServlet extends HttpServlet {
 	 */
 	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 		
+		Member logined = (Member) request.getSession().getAttribute("logined");
 		
+		if (logined != null) {
+			if(logined.getIsAdmin() == 1){
 		
-		String id =request.getParameter("del");	
-		
-		String [] arr=id.split(",");
-		
-		
-		int result=new AdminService().deleteBook(arr);
-		
-		String msg="";
-		String view="/views/common/msg.jsp";
-		String loc="";
-		
-		if(result==arr.length) {
-			msg="삭제 성공";
-			loc="/admin/book";
+				String id =request.getParameter("del");	
+				
+				String [] arr=id.split(",");
+				
+				
+				int result=new AdminService().deleteBook(arr);
+				
+				String msg="";
+				String view="/views/common/msg.jsp";
+				String loc="";
+				
+				if(result==arr.length) {
+					msg="삭제 성공";
+					loc="/admin/book";
+				}else {
+					msg="삭제실패";
+					loc="/admin/book";
+				}
+				
+				
+					
+				
+				request.setAttribute("msg", msg);
+				request.setAttribute("loc", loc);
+				request.getRequestDispatcher(view).forward(request, response);
+			}else {
+				//관리자가 아닐 때
+				request.setAttribute("msg", "접근할 수 없는 페이지입니다.");
+				request.setAttribute("loc", "/main/mainView");
+				request.getRequestDispatcher("/views/common/msg.jsp").forward(request, response);;
+			}
 		}else {
-			msg="삭제실패";
-			loc="/admin/book";
+			//로그인을 안했을 때
+			response.sendRedirect(request.getContextPath()+"/member/login");
 		}
-		
-		
-			
-		
-		request.setAttribute("msg", msg);
-		request.setAttribute("loc", loc);
-		request.getRequestDispatcher(view).forward(request, response);
 		
 	}
